code refactoring of easyform adapter
