+++ src/collective/ifttt/actions/configure.zcml
@@ -44,8 +44,4 @@
         factory=".ifttt.IftttTriggerAction"
         />
 
-    <!-- IFTTT adapter for easyform -->
-
-    <include file="actions.zcml" />
-
 </configure>
