+++ src/collective/ifttt/profiles/default/metadata.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <metadata>
-  <version>1011</version>
+  <version>1012</version>
   <dependencies>
     <!--<dependency>profile-plone.app.dexterity:default</dependency></dependencies>-->
   </dependencies>
