+++ src/collective/ifttt/upgrades.zcml
@@ -4,7 +4,7 @@
 
   <genericsetup:upgradeStep
     source="*"
-    destination="1011"
+    destination="1012"
     title="Reload GS profiles"
     description=""
     profile="collective.ifttt:default"
